requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.3.1
attrs==19.3.0
billiard==3.6.3.0
black==19.10b0
celery==4.4.2
click==7.1.2
dateparser==1.0.0
dj-database-url==0.5.0
Django==3.1.4
django-autoslug==1.9.7
django-cors-headers==3.4.0
django-crispy-forms==1.10.0
django-fav==0.3
django-favit==0.4.0
django-favorite==0.0.1
django-mixins==0.0.10
django-search-views==0.3.7
django-taggit==1.3.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
docutils==0.16
flake8==3.8.2
Flask==1.1.2
generic==1.0.0
gunicorn==20.0.4
importlib-metadata==3.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.11
lockfile==0.12.2
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mixins==0.1.4
mock==4.0.2
note==0.5.2
pathspec==0.8.0
Pillow==8.0.1
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pymongo==3.11.1
python-daemon==2.2.4
python-dateutil==2.8.1
pytz==2020.4
pyzmq==20.0.0
redis==3.5.2
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
tornado==6.1
typed-ast==1.4.1
tzlocal==2.1
vine==1.3.0
Werkzeug==1.0.1
whitenoise==5.2.0
zipp==3.4.0
